Added a very basic wall to the scene for testing.
Created a Shooter class as a parent for all ranged characters :).
