+++ KaiFighter/KaiFighterGame/Content/Content.mgcb
@@ -49,3 +49,15 @@
 /processorParam:TextureFormat=Color
 /build:Images/Projectiles/1.png
 
+#begin Images/Statics/line.png
+/importer:TextureImporter
+/processor:TextureProcessor
+/processorParam:ColorKeyColor=255,0,255,255
+/processorParam:ColorKeyEnabled=True
+/processorParam:GenerateMipmaps=False
+/processorParam:PremultiplyAlpha=True
+/processorParam:ResizeToPowerOfTwo=False
+/processorParam:MakeSquare=False
+/processorParam:TextureFormat=Color
+/build:Images/Statics/line.png
+
