+++ KaiFighter/KaiFighterGame/Content/Images/Statics/line.png
